requirements.txt
# Runtime
discord.py[voice]==2.4.0
PyNaCl==1.5.0
python-dotenv==1.0.1
aiohttp==3.10.5
aiosqlite==0.20.0
pydantic==2.8.2

# Dev (optional)
# uvloop temporarily removed due to Python 3.13 build issues